requirements.txt
torch
pandas
PyPDF2
fastapi
uvicorn
tiktoken
langchain
python-dotenv
huggingface-hub
langchain-openai
langchain-community
langchain-huggingface
sentence-transformers